www/firefox: now really use ccache
The problem is the system includes bsd.ccache.mk after bsd.overlay.mk. So at
this moment we have empty CCACHE_BIN. But we need to know it, because with just
"--with-ccache" it can't find ccache. So I've just copied a snipped from
bsd.ccache.mk, which sets CCACHE_BIN.